requirements.txt
rootpy
root-numpy
numpy
matplotlib
nose
sphinx
pytables
uncertainties